refactor: presentation 레이어 정리
