--- src/main/java/com/techfork/useraccount/dto/UserInterestResponse.java
@@ -1,11 +1,0 @@
-package com.techfork.useraccount.dto;
-
-import lombok.Builder;
-
-import java.util.List;
-
-@Builder
-public record UserInterestResponse(
-        List<UserInterestDto> interests
-) {
-}
