+++ src/main/java/com/techfork/useraccount/presentation/UserConverter.java
@@ -1,12 +1,12 @@
-package com.techfork.useraccount.converter;
+package com.techfork.useraccount.presentation;
 
 import com.techfork.useraccount.application.command.CompleteOnboardingCommand;
 import com.techfork.useraccount.application.command.UpdateAccountProfileCommand;
 import com.techfork.useraccount.application.command.UserInterestCommand;
 import com.techfork.useraccount.application.query.result.GetAccountProfileResult;
-import com.techfork.useraccount.dto.AccountProfileResponse;
-import com.techfork.useraccount.dto.OnboardingRequest;
-import com.techfork.useraccount.dto.UpdateAccountProfileRequest;
+import com.techfork.useraccount.presentation.request.OnboardingRequest;
+import com.techfork.useraccount.presentation.request.UpdateAccountProfileRequest;
+import com.techfork.useraccount.presentation.response.AccountProfileResponse;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
