+++ src/main/java/com/techfork/useraccount/presentation/request/SaveInterestRequest.java
@@ -1,4 +1,4 @@
-package com.techfork.useraccount.dto;
+package com.techfork.useraccount.presentation.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
@@ -8,6 +8,6 @@ import java.util.List;
 public record SaveInterestRequest(
         @NotNull(message = "관심사 목록은 필수입니다.")
         @NotEmpty(message = "관심사를 최소 1개 이상 선택해주세요.")
-        List<UserInterestDto> interests
+        List<UserInterestRequest> interests
 ) {
 }
