+++ src/main/java/com/techfork/useraccount/presentation/response/UserInterestResponse.java
@@ -1,0 +1,17 @@
+package com.techfork.useraccount.presentation.response;
+
+import lombok.Builder;
+
+import java.util.List;
+
+@Builder
+public record UserInterestResponse(
+        List<Interest> interests
+) {
+    @Builder
+    public record Interest(
+            String category,
+            List<String> keywords
+    ) {
+    }
+}
